Use latest YAML and QuoteNumericStrings feature.
This is to ensure that strings that resemble numeric keys are encoded
quoted as strings such that decoders correctly decode them as strings,
not numbers.

Thus, trailing 0's ( which are significant ) are not lost,
and there's no subsequent

  Perl: 5.1
  Perl: 5.2

Craziness in your travis smoke list.

( Which are perls version 5.10 and 5.20 respectively )